jolie foret de if .... :/ - faire lookup table                  [todo] - faire fonction pour save les element[todo]

+++ srcs/main.cpp
@@ -6,43 +6,46 @@
 /*   By: bcano <bcano@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/07 12:19:46 by bcano             #+#    #+#             */
-/*   Updated: 2022/06/19 19:14:19 by anclarma         ###   ########.fr       */
+/*   Updated: 2022/06/21 21:57:53 by anclarma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <iostream>
-#include <cstring>
-#include <cstdlib>
-#include <ctype.h>
 #include "Server.hpp"
+#include <cstdlib>
+#include <cstring>
+#include <ctype.h>
+#include <iostream>
 
-int    display_error(int type, const char *err, const char *message)
+int display_error(int type, const char *err, const char *message)
 {
-    std::cout << "./ircserv: error: " << err << std::endl;
-    std::cout << message << std::endl;
-    return (type);
+	std::cout << "./ircserv: error: " << err << std::endl;
+	std::cout << message << std::endl;
+	return (type);
 }
 
 int main(int argc, char **argv)
 {
-    int port = 0;
-    std::string pwd("");
-    
-    if (argc != 3)
-        return (display_error(1, "Wrong numbers of Parameters","Usage: ./ircserv <port> <password>"));
-    for (int i ( 0 ); argv[1][i]; i++)
-        if (!(isdigit(argv[1][i])))
-            return (display_error(1, "Port must be a number", "Usage: ./ircserv <port> <password>"));
-    if (atoi(argv[1]) <= 0)
-        return (display_error(1, "Port must be strictly number", "Usage: ./ircserv <port> <password>"));
-    port = atoi(argv[1]);
-    pwd = argv[2];
-    std::cout << port << " " << pwd << std::endl;
-	Server	s(port, pwd);
+	uint16_t	port = 0;
+	std::string	pwd("");
+
+	if (argc != 3)
+		return (display_error(1, "Wrong numbers of Parameters",
+							  "Usage: ./ircserv <port> <password>"));
+	for (int i(0); argv[1][i]; i++)
+		if (!(isdigit(argv[1][i])))
+			return (display_error(1, "Port must be a number",
+								  "Usage: ./ircserv <port> <password>"));
+	if (atoi(argv[1]) <= 0)
+		return (display_error(1, "Port must be strictly number",
+							  "Usage: ./ircserv <port> <password>"));
+	port = static_cast<uint16_t>(atoi(argv[1]));
+	pwd = argv[2];
+	std::cout << port << " " << pwd << std::endl;
+	Server s(port, pwd);
 	s.create_sock();
 	s.set_sock();
 	s.bind_sock();
 	s.listen();
 	s.poll_loop();
-    return (0);
+	return (0);
 }
